requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.2